Add Pay Now button to Meet Entry List.

+++ src/app/pending-entry-action/pending-entry-action.component.ts
@@ -6,6 +6,7 @@ import {ActivatedRoute, ActivatedRouteSnapshot, Router} from '@angular/router';
 import {MeetService} from '../meet.service';
 import {MemberService} from '../member.service';
 import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
+import {FormBuilder} from '@angular/forms';
 
 @Component({
   selector: 'app-pending-entry-action',
@@ -22,6 +23,7 @@ export class PendingEntryActionComponent implements OnInit {
   meetFee = 0;
 
   entry;
+  incompleteEntry;
   eventEntries: EntryEvent[];
 
   statusLabel = '';
@@ -46,53 +48,64 @@ export class PendingEntryActionComponent implements OnInit {
   canBeActioned = false;
   existingEntryShow = false;
 
+  editing = false;
+  pendingActionForm;
+
   constructor(private entryService: EntryService,
               private meetService: MeetService,
               private memberService: MemberService,
               private route: ActivatedRoute,
               private router: Router,
-              private modalService: NgbModal) { }
+              private modalService: NgbModal,
+              private fb: FormBuilder) { }
 
   ngOnInit() {
+
+    this.pendingActionForm = this.fb.group({
+      entrantSurname: '',
+      entrantFirstName: '',
+      entrantDob: '',
+      entrantGender: '',
+      entrantNumber: '',
+      entrantClub: ''
+    });
+
     this.pendingEntryId = this.route.snapshot.paramMap.get('pendingId');
     this.entryService.getPendingEntry(this.pendingEntryId).subscribe((entry: any) => {
-      this.entry = entry.entrydata;
-      console.log(entry);
-      this.meet_id = entry.meet_id;
-      this.meet = this.meetService.getMeet(this.meet_id);
-      console.log(this.meet);
-      this.meetName = this.meet.meetname;
-      this.meetFee = this.meet.meetfee;
-      this.eventEntries = entry.entrydata.entryEvents;
-      this.statusLabel = entry.status_label;
-      this.statusText = entry.status_description;
-      if (this.entry.paymentDetails !== undefined && this.entry.paymentDetails !== null) {
-        this.paidAmount = this.entry.paymentDetails.amount;
-      }
-      this.pendingReason = entry.pending_reason;
-      this.created_at = entry.created_at;
-      this.updated_at = entry.updated_at;
+      this.loadEntry(entry);
+    });
+  }
 
-      this.memberService.getMemberByNumber(this.entry.membershipDetails.member_number).subscribe((member: any) => {
-        console.log(member);
-        this.memberSearchResult = member.member;
+  loadEntry(entry) {
+    this.incompleteEntry = entry;
+    this.entry = this.incompleteEntry.entrydata;
+    console.log(entry);
+    this.meet_id = entry.meet_id;
+    this.meet = this.meetService.getMeet(this.meet_id);
+    this.meetName = this.meet.meetname;
+    this.meetFee = this.meet.meetfee;
+    this.eventEntries = entry.entrydata.entryEvents;
+    this.statusLabel = entry.status_label;
+    this.statusText = entry.status_description;
+    if (this.entry.paymentDetails !== undefined && this.entry.paymentDetails !== null) {
+      this.paidAmount = this.entry.paymentDetails.amount;
+    }
+    this.pendingReason = entry.pending_reason;
+    this.created_at = entry.created_at;
+    this.updated_at = entry.updated_at;
 
-        this.entryService.getSubmittedEntriesByMemberNumber(this.entry.membershipDetails.member_number, this.meet_id)
-          .subscribe((entries: any) => {
-            this.existingEntries = entries;
-            console.log(this.existingEntries);
-            if (this.existingEntries !== undefined && this.existingEntries !== null) {
-              if (this.existingEntries.length > 0) {
-                this.existingEntryShow = true;
-                this.canBeActioned = false;
-              } else {
-                if (this.memberSearchResult.number === this.entry.membershipDetails.member_number) {
-                  this.canBeActioned = true;
-                } else {
-                  this.canBeActioned = false;
-                  console.log('Membership number does not match');
-                }
-              }
+    this.memberService.getMemberByNumber(this.entry.membershipDetails.member_number).subscribe((member: any) => {
+      console.log(member);
+      this.memberSearchResult = member.member;
+
+      this.entryService.getSubmittedEntriesByMemberNumber(this.entry.membershipDetails.member_number, this.meet_id)
+        .subscribe((entries: any) => {
+          this.existingEntries = entries;
+          console.log(this.existingEntries);
+          if (this.existingEntries !== undefined && this.existingEntries !== null) {
+            if (this.existingEntries.length > 0) {
+              this.existingEntryShow = true;
+              this.canBeActioned = false;
             } else {
               if (this.memberSearchResult.number === this.entry.membershipDetails.member_number) {
                 this.canBeActioned = true;
@@ -101,9 +114,18 @@ export class PendingEntryActionComponent implements OnInit {
                 console.log('Membership number does not match');
               }
             }
-          });
-      });
-
+          } else {
+            if (this.memberSearchResult.number === this.entry.membershipDetails.member_number) {
+              this.canBeActioned = true;
+            } else {
+              this.canBeActioned = false;
+              console.log('Membership number does not match');
+            }
+          }
+        });
+    }, (error: any) => {
+      console.log('unable to find member');
+      this.memberSearchResult = null;
     });
   }
 
@@ -151,15 +173,47 @@ export class PendingEntryActionComponent implements OnInit {
 
   delete() {
     this.modalService.open(this.deleteConfirm, {ariaLabelledBy: 'modal-basic-title'}).result.then((result) => {
-      console.log(result)
-      this.entryService.deletePending(this.pendingEntryId).subscribe((deleted: any) => {
-        console.log(deleted);
-        this.router.navigate(['/', 'pending-entries', this.meet_id]);
-      }, (error: any) => {
-        console.log(error);
-      });
+      console.log(result);
+      if (result === 'Yes') {
+        this.entryService.deletePending(this.pendingEntryId).subscribe((deleted: any) => {
+          console.log(deleted);
+          this.router.navigate(['/', 'pending-entries', this.meet_id]);
+        }, (error: any) => {
+          console.log(error);
+        });
+      }
     }, (reason) => {
       console.log(reason)
     });
+  }
+
+  edit() {
+    this.pendingActionForm.patchValue({
+      entrantSurname: this.entry.entrantDetails.entrantSurname,
+      entrantFirstName: this.entry.entrantDetails.entrantFirstName,
+      entrantGender: this.entry.entrantDetails.entrantGender,
+      entrantDob: this.entry.entrantDetails.entrantDob,
+      entrantNumber: this.entry.membershipDetails.member_number
+    });
+
+    this.editing = true;
+  }
+
+  update() {
+    console.log(this.entry);
+    this.editing = false;
+
+    this.entry.entrantDetails.entrantSurname = this.pendingActionForm.controls['entrantSurname'].value;
+    this.entry.entrantDetails.entrantFirstName = this.pendingActionForm.controls['entrantFirstName'].value;
+    this.entry.entrantDetails.entrantGender = this.pendingActionForm.controls['entrantGender'].value;
+    this.entry.entrantDetails.entrantDob = this.pendingActionForm.controls['entrantDob'].value;
+    this.entry.membershipDetails.member_number = this.pendingActionForm.controls['entrantNumber'].value;
+    this.entryService.updatePending(this.pendingEntryId, this.incompleteEntry).subscribe((entry) => {
+      this.loadEntry(entry);
+    });
+  }
+
+  cancel() {
+    this.editing = false;
   }
 }
